requirements.txt
requests
fastapi
pydantic-settings
python-dotenv